requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2    
async-timeout==3.0.1
attrs==21.4.0       
chardet==4.0.0      
colorama==0.4.4     
Deprecated==1.2.13
discord.py==1.7.3
future==0.18.2
idna==3.3
multidict==6.0.2
packaging==21.3
pefile==2021.9.3
pyarmor==7.3.6
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.1
pylock==0.3
pyparsing==3.0.7
pywin32-ctypes==0.2.0
redis==4.1.3
typing_extensions==4.1.0
wrapt==1.13.3
yarl==1.7.2